Integrated TableWithMeatballMenu with Table
