+++ src/components/v5/common/ActionSidebar/partials/BatchPaymentsTable/hooks.tsx
@@ -1,8 +1,7 @@
 import { createColumnHelper, type ColumnDef } from '@tanstack/react-table';
-import { useCallback, useMemo } from 'react';
+import { useMemo } from 'react';
 
 import { formatText } from '~utils/intl.ts';
-import { type TableWithMeatballMenuProps } from '~v5/common/TableWithMeatballMenu/types.ts';
 
 import { type BatchPaymentsTableModel } from './types.ts';
 
@@ -44,21 +43,3 @@ export const useBatchPaymentsTableColumns = (): ColumnDef<
 
   return columns;
 };
-
-export const useGetTableMenuProps = ({ remove }) =>
-  useCallback<
-    TableWithMeatballMenuProps<BatchPaymentsTableModel>['getMenuProps']
-  >(
-    ({ index }) => ({
-      cardClassName: 'min-w-[9.625rem] whitespace-nowrap',
-      items: [
-        {
-          key: 'remove',
-          onClick: () => remove(index),
-          label: formatText({ id: 'table.row.remove' }),
-          icon: 'trash',
-        },
-      ],
-    }),
-    [remove],
-  );
