--- src/components/v5/common/TableWithMeatballMenu/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TableWithMeatballMenu.tsx';
