--- src/components/v5/common/TableWithMeatballMenu/types.ts
@@ -1,11 +1,0 @@
-import { type Row } from '@tanstack/react-table';
-
-import { type MeatBallMenuProps } from '~v5/shared/MeatBallMenu/types.ts';
-
-import { type TableProps } from '../Table/types.ts';
-
-export interface TableWithMeatballMenuProps<T> extends TableProps<T> {
-  getMenuProps: (row: Row<T>) => MeatBallMenuProps | undefined;
-  meatBallMenuSize?: number;
-  meatBallMenuStaticSize?: string;
-}
